Refactor WebView cleanup demo and test for clarity and completeness

+++ demos/android/MASVS-PLATFORM/MASTG-DEMO-00de/script.js
@@ -1,24 +1,78 @@
 function enumerateDeleteAllDataMethod() {
-  return Java.enumerateMethods('com.android.webview.chromium.*!deleteAllData')[0]
+  const res = Java.enumerateMethods('com.android.webview.chromium.*!deleteAllData');
+  return res && res[0];
 }
 
-Java.perform(() => {
-  console.log('Enumerating chromium...');
+function enumerateSetDomStorageEnabledMethod() {
+  const res = Java.enumerateMethods('com.android.webview.chromium.*!setDomStorageEnabled');
+  return res && res[0];
+}
+
+function ensureClassLoaded(className) {
+  try {
+    Java.use(className);
+  } catch (e) {
+  }
+}
+
+function printBacktrace(maxLines = 8) {
+  let Exception = Java.use("java.lang.Exception");
+  let stackTrace = Exception.$new().getStackTrace().toString().split(",");
+
+  console.log("\nBacktrace:");
+  for (let i = 0; i < Math.min(maxLines, stackTrace.length); i++) {
+      console.log(stackTrace[i]);
+  }
+  console.log("\n");
+}
+
+Java.perform(function () {
+  console.log('Enumerating chromium.');
 
   if (enumerateDeleteAllDataMethod() === undefined) {
-    console.log('Bring WebStorage to memory so we can hook its deleteAllData() method.');
-    // If WebStorage is not yet initialized, bring it to memory.
-    Java.use("android.webkit.WebStorage").getInstance()
+    console.log('Bring WebStorage to memory so we can hook its deleteAllData method.');
+    Java.use('android.webkit.WebStorage').getInstance();
   }
 
-  const method = enumerateDeleteAllDataMethod()
-  Java.classFactory.loader = method.loader;
+  const deleteAllDataMethod = enumerateDeleteAllDataMethod();
+  if (deleteAllDataMethod !== undefined) {
+    Java.classFactory.loader = deleteAllDataMethod.loader;
+    const WebStorageAdapter = Java.use(deleteAllDataMethod.classes[0].name);
 
-  const WebStorageAdapter = Java.use(method.classes[0].name);
-  WebStorageAdapter.deleteAllData.implementation = function () {
-    console.log('WebStorage.deleteAllData() called');
-    return this.deleteAllData();
+    WebStorageAdapter.deleteAllData.implementation = function () {
+      console.log('WebStorage.deleteAllData called.');
+      printBacktrace();
+      return this.deleteAllData();
+    };
+
+    console.log('WebStorage.deleteAllData hooked.');
+  } else {
+    console.log('WebStorage.deleteAllData not found.');
   }
 
-  console.log('WebStorage.deleteAllData() hooked.');
+  ensureClassLoaded('android.webkit.WebView');
+  ensureClassLoaded('android.webkit.WebSettings');
+
+  const domMethod = enumerateSetDomStorageEnabledMethod();
+  if (domMethod !== undefined) {
+    Java.classFactory.loader = domMethod.loader;
+    const ContentSettingsAdapter = Java.use(domMethod.classes[0].name);
+
+    if (ContentSettingsAdapter.setDomStorageEnabled.overloads.length === 0) {
+      console.log('setDomStorageEnabled found but no overloads exposed.');
+    } else {
+      ContentSettingsAdapter.setDomStorageEnabled.overloads.forEach(function (ov) {
+        ov.implementation = function () {
+          const enabled = arguments.length > 0 ? arguments[0] : undefined;
+          console.log('ContentSettingsAdapter.setDomStorageEnabled called, enabled is ' + enabled + '.');
+          printBacktrace();
+          return ov.apply(this, arguments);
+        };
+      });
+
+      console.log('ContentSettingsAdapter.setDomStorageEnabled hooked.');
+    }
+  } else {
+    console.log('ContentSettingsAdapter.setDomStorageEnabled not found.');
+  }
 });
